requirements.txt
requests==2.31.0
pandas==2.2.0
python-dotenv==1.0.0
scikit-learn==1.4.0
replicate==0.22.0 